+++ changelog.md
@@ -1,2 +1,9 @@
 # V1.0.0
 Initial release.
+
+# V1.1.0
+
+
+patch: fixes issue where in Dwcom would not recognize that speechInterrupt was set to false.
+patch: Makes speach output say left/joined channel root rather than left/joined channel the root channel.
+feat: Allows the use of logins and logouts txt files to create random login and logout messages.